fixed missing await in initializeProxyHandler
fixed missing await causing proxy initialization after usage
moved Global-/PerSiteStrictMode to utilities.ts just in case to prevent circular dependencies and prevent dependency on background.ts
