+++ core/background.ts
@@ -7,29 +7,19 @@ import "./vendor/browser-polyfill.js";
 
 import {initializeProxyHandler, loadProxySettings} from "./background_helpers/proxy_handler.js";
 import {allowAllgeofence, geofence, resetPolicyCookie} from "./background_helpers/geofence_handler.js";
-import {getSyncValue, GLOBAL_STRICT_MODE, ISD_ALL, ISD_WHITELIST, PER_SITE_STRICT_MODE, saveSyncValue, type SyncValueSchema} from "./shared/storage.js";
+import {GLOBAL_STRICT_MODE, ISD_ALL, ISD_WHITELIST, PER_SITE_STRICT_MODE, type SyncValueSchema} from "./shared/storage.js";
 import {globalStrictModeUpdated, initializeDnr, perSiteStrictModeUpdated, updateProxySettingsInDnrRules} from "./background_helpers/dnr_handler.js";
 import {initializeRequestInterceptionListeners} from "./background_helpers/request_interception_handler.js";
 import {initializeTabListeners} from "./background_helpers/tab_handler.js";
-import {initializeIsChromium} from "./shared/utilities.js";
+import {initializeIsChromium, initializeStrictModes, setGlobalStrictMode, setPerSiteStrictMode} from "./shared/utilities.js";
 
-export let GlobalStrictMode: SyncValueSchema[typeof GLOBAL_STRICT_MODE] = false;
-export let PerSiteStrictMode: SyncValueSchema[typeof PER_SITE_STRICT_MODE] = {};
 
 /*--- setup ------------------------------------------------------------------*/
 // Initialize IsChromium
 initializeIsChromium();
 
 const initializeExtension = async () => {
-    const storageGlobalStrictMode = await getSyncValue(GLOBAL_STRICT_MODE);
-    GlobalStrictMode = storageGlobalStrictMode ?? false;
-    if (storageGlobalStrictMode === undefined) await saveSyncValue(GLOBAL_STRICT_MODE, GlobalStrictMode);
-    console.log("[initializeExtension]: GlobalStrictMode:", GlobalStrictMode);
-
-    const storagePerSiteStrictMode = await getSyncValue(PER_SITE_STRICT_MODE);
-    PerSiteStrictMode = storagePerSiteStrictMode ?? {};
-    if (storagePerSiteStrictMode === undefined) await saveSyncValue(PER_SITE_STRICT_MODE, PerSiteStrictMode);
-    console.log("[initializeExtension]: PerSiteStrictMode:", PerSiteStrictMode);
+    await initializeStrictModes();
 
     /*--- PAC --------------------------------------------------------------------*/
     // initializing proxy handler before DNR, as some DNR rules rely on the `proxyAddress`
@@ -59,14 +49,14 @@ browser.storage.onChanged.addListener(async (changes, namespace) => {
 
         } else if (changes.perSiteStrictMode?.newValue !== undefined) {
 
-            PerSiteStrictMode = (changes.perSiteStrictMode.newValue || {}) as SyncValueSchema[typeof PER_SITE_STRICT_MODE];
+            setPerSiteStrictMode((changes.perSiteStrictMode.newValue || {}) as SyncValueSchema[typeof PER_SITE_STRICT_MODE]);
 
             // update DNR rules
             await perSiteStrictModeUpdated();
 
         } else if (changes.globalStrictMode?.newValue !== undefined) {
 
-            GlobalStrictMode = changes.globalStrictMode.newValue as SyncValueSchema[typeof GLOBAL_STRICT_MODE];
+            setGlobalStrictMode(changes.globalStrictMode.newValue as SyncValueSchema[typeof GLOBAL_STRICT_MODE]);
 
             // update DNR rules
             await globalStrictModeUpdated();
